Store vault/deposit metadata in bags
Initial (simple) serialisation of objects to JSON. The vault/deposit
information is stored in a new top-level "metadata" directory in the
bag.

In future we should probably use a RDF or JSON-LD representation of
metadata (perhaps use the ORE schema/fields?).

Tag metadata (e.g. md5 fixity) also needs to be added for these new
files.