Do not persist Job ID string from worker in Event
There are currently two instances of the Job ID in the Event class. One
represents an object relationship with a Job and is persisted in
Hibernate. The other is used at initial creation time by the worker
generating a stream of events (which has no DB access). These generated
events are sent to the broker which is responsible for persistence.