Implement different verification strategies
Default COPY_BACK method to restore data from archive storage, extract
and validate the bag.

Alternative LOCAL_ONLY method for services where a copy back is
undesirable (e.g. glacier where charges would apply and data is not
immediately available). In this case the temporary files are extracted
and validated but the copy to the archive system is not fully checked.

Additional checks of the file hash could be added later (but only where
the archive storage system exposes this information).