Set concurrency for broker event listener
By default rabbitmq event listeners are spun up depending on demand.
This means that an event later in the queue can overtake an earlier
event (e.g. while it is waiting for a database query).

For now, I’ve set the event listener to process events in order by
limiting the concurrency setting of the listener-container. If we want
concurrent event processing we’ll need to take special care that the
event handling code is thread-safe (especially with regard to
hibernate/database updates).