IssueId #399 Finally added some unit tests for the TSM plugin
1) Added tests to TivoliStorageManagerTest.java for the store method.  One checks that storing works by storing / querying then deleting the archive the other just checks that trying to store a non existent file throws the expected exception.
2) There are also some empty placeholders for other tests these will be filled in as I progress.

I should be able to quickly add tests for retrieve shortly.

One thing to keep in mind when either checking out this branch or especially after this is merged to master is that you will need the TSM client installed for it to work how that affects Manchester and Jisc I dunno.

We’ll also have to make sure the tests work in William’s travis / docker builds