IssueId RSS022-112 - Add chunking Step
It's now possible to turn on chunking by changing the new property:
`chunking.enabled = true`

This will chunk the deposit tar file into 500MB chunks (this is the
current default chunk size but I'm planning to make this cinfigurable
too).

I had to change the Deposit table to add the number of chunks created so
the data can be retrieved.