IssueId RSS022-112 - Fix integration tests (again)
I forgot to change the expected file name, as it's being chunk now it
will have a tailing '.1'
Tests passed on my machine because I already had the expected tar file,
created when run the test in the pass.

I also made sure that the file gets deleted at the end of the test so it
doesn't happen again.