Issue # - Handle checksum for chunk file
The deposit will now generate a checksum for each chunks and save it
into the database.
When retrieving the data, each chunk will be compare
with their expected checksum.
The integration test WorkerInstanceIT have been update to also test the
retrieving of data in order to test.
The database will now contain a new table "DepositChunk" which will
mainly contain each chunk checksum.