RSS022-101 - Add checksum for encrypted file
A checksum of the encrypted file is now kept in the database, that will
allow to determine if any checksum fail where the problem is comming
from: encryption, taring, chunking...

I also had to fix some issue with Deposit and Retrieval as it wasn't
working anymore if encryption or chunking was turned off. I think we
fail to see these issues when we've merged the TSM and chunkingi work
together.