IssueId #RSS022-126 Added values to datavault.properties for the TSM and S3 plugins
1) Added optionsDir and bucketName params to the datavault.properties file if these params are missing and the plugins are selected for use default values from the plugins are used.
2) Added new addArchiveSpecificOptions method to the DepositsController and updated the deposit and retrieve method to use it.
3) Updated the datavault-broker-servlet.xml config file to inject the new values from the properties file to the deposits controller
4) Updated the S3Cloud and TivoliStorageManager plugins to use the new values if available

I've tested this manually with correct values in the properties, bad properties, missing properties etc. it all seems to work.

I might try and inject the values directly into the plugin if I get a chance but I'm not sure that is going to work the way things are set up.  If I get time I'll try.