Issue #RSS022-139 - Add University Header and Footer
Add Jackie header and footer to the current pages.
I had to hack the CSS to get it to work properly as some of the CSS were
conflicting.
There is a jackie-quickfix.css which basically contains what was in
vault.css minus the conflicting part, plus a few trick to keep the
footer at the bottom.