IssueId #RSS022-78 Tidy up off the PURE integration code
1) Updated ExternalMetadataService so that strings were moved to conts and handling was added for various requests returning null,  I also reverted to using the original MockProvider and replaced the PureProvider with my new one.  This should mean that everything will continue to work as it did before my changes unless you turn on PURE as the metadate provider in the config which will now use my new TestPureProvider.
2) Removed a couple of unused imports from DatabaseAuthenticationProvider
3) Removed an unused import from VaultController
4) Added some logging to the LDAPService this can probably be removed again once we have tested things that actually use the LDAP version of this code.

I've not managed to test this on my machine yet, I'm probably going to give up on that and just test straight on test.datavault.  I spent a couple of days trying to throw something together that would test on my machine,  I've added comments about that here https://www.jira.is.ed.ac.uk/browse/RSS022-78.

The gist is that if you don't turn on PURE as the external metadata provider the current mock data is used, if you do it will try and get the LDAP attribs from the Users table but if they are missing (they will be) your employee id will be hardcode to one that gets datasets from the flat file.

Even testing on test.datavault is going to be a little tricky as an existing config bug means that everyone that has logged in to test so far has missing LDAP attribs I think if I delete the users they should be added when they are readded to the db.

I should be able to set this up so people can look at it at the next sprint meeting to decide whether it is ready to go into master and onto test.datavault.