IssueId #RSS022-181 Potential fix for the TSM retrieve issue
This is a potential fix for the TSM issue where using a different worker from the one used to deposit causes the retrieve to fail.

I've updated the TSM plugin so that each file or chunk (depending on whether chunking is turned on) is now moved to a separate temp space while it is deposited / retrieved to / from each node.  This will slow things down but it does work on my machine, I can still retrieve after I've stopped and started the worker and Tomcat / worker too.

I'm not sure if this solution will work without further changes when we get around to threading the storage of the 3 copies, it might be better if we just add the original temp file location to the DB so that it can be extracted for the TSM plugin.

I'm committing this now so it doesn't get lost but I'm not going to merge until we've discussed at a sprint meeting.