Issue RSS022-79 - Integration of HarshiCorp Vault
Added container for running a Vault server and one to set it up i.e.
'vault-administration'
Added other modification to the dockerisation

Added a Maven dependency: vault-java-driver

Change the Worker so it gets the secret key for encryption from the
Harshcorp Vault.
Hashicorp Vault information to retrieve the key are defined in the
datavault.properties file.