Issue RSS022-203 - Encrypt private ssh key
Encrypt key before inserting it in the database.
Decrypt key when trying to access SFTP storage
Move all Encryption code to common to make it accessible from every module:
worker, broker.
Redifine the FileStore properties as a HashMap<String, Object> in order
to be able to save encrypted key as a byte[], as the conversion to String
was causing some issues.