Issue RSS022-203 - Fix SSH key encryption
I didn't test the whole process before creating the merge, my mistake.
Because we're sending the key to RabbitMQ as a JSON string the key will
have to be converted to a String.
I found out that if I'm using the Base64 object made by BouncyCastle
this seems to work.
So I put back the FileStore properties as HashMap<String,String>.