RSS022-157 - Removed date format isntruction
It was more confusing than helping as when browser have a date picker
they don't especially display the date in this format, so instead I used
placeholder which should only display if there is no date picker.

Safari actually have a date-picker but it's really bad and let you enter
numbers but at least the jquery.validate() should flag it as an error if
it happens.