Fix date on Create Vault page
It's now possible to create a vault without giving a Grant End Date.
I also change the regular expression so it only accept 20XX years and a proper month day.
Ofcourse this is not as good as a proper date picker but we'll have to
find one which is working on every browser.