IssueId #RSS022-251 Add option so that if chunking in on we process as much as possible on local disk
Our test, demo and live machines are set up to use a mounted disk comprising of  datastore disc,  this disk is slower than local disk or disk on the san we have also had complaints that our use of this disk is putting unacceptable strain on the whole of datastore.

Due to this we have added an option that allows the configuration of 'local' disk space to be used for processing a chunk (in our case the local disk is actual SAN space assigned to a VM).

This is done by the addition of the new datavault.properties value localTempDir.  If there is a value for this option and it is different to the tempDir option everything apart from the copy of the deposit files from the user's datastore to our datastore and the recomposition of the retrieved files during verification will be carried out on the localTempDir space.

I've made the following changes to the Deposit option so far to implement it, nothing has been changed for Retrieve at present.

1) Updated docker and default datavault.properties to have a new localTempDir param
2) Updated docker-compose.yml to create a new dir to act as the localTemDir value for the worker
3) Updated datavault-broker-servlet.xml and datavault-worker.xml to pick up the new localTempDir
4) Updated DepositsController to have a member and setter for the localTempDir option
5) Updated Context to have a new localTempDir member along with accessors and an updated constructor
6) Updated Receiver to have a member and setter for the localTempDir option it also uses the member to create the dir if tempDir and localTempDir are different and adds the new value to the context
7) Updated Deposit so that whether chunking is enabled or not we use chunking code ( by adding 1 chunk if chunking is turned off).  I also refactored so that after the file is copied over and split (or not) each chunk is individually moved to local disk (if the local disk is specified and different to the default) then encrypted, checksumed etc.  The validation has also been refactored so that if local disk is configured it is used to extract and decrypt before copying the chunk to datastore for recomposition and final checks.