IssueId #RSS022-259 Updated the location of the PURE flat files to default to /tmp/ but be configurable.
I've updated the PURE flat file code so that the dir the code expects to find the flat files has changed.  It defaults to use /tmp as before but if datavault.properties contains a value for flatfile.location it will use that instead.
I'll need to update the flat file generation scripts to pick this up somehow too and we'll need to remember to change it on the demo / live deploys away from /tmp

1) Updated datavault-broker-root.xml so that the ExternalMetadatService now injects the flatfile.location value from the datavault.properties file (it defaults to /tmp if missing)
2) Updated ExternalMetadataService.java so that in now has a flatFileLocation member along with a setter.  I also updated setMetadataURL to use the new member to set the value into PureFlatFileProvider
3) Updated PureFlatFileProvider.java so that it has new members for the configurable file location along with gettters and setters