RSS022-305 - Make sure Deposit and Retrieve stop if any Exception is thrown
We had previous issue where the Deposit would continue taring,
encrypting, archiving even though the File transfer failed in the middle
of the process. We believe this is because the catch close doesn't
contain a `return`.o I looks at every catch close which could have the
same issue in Deposit.java and Retrieve.java and added the `return`.