RSS022-310 - Retrieve in Timestamp folder
I first cretaed a timestamped folder to avoid overwriting files as we
though the error we were experienting were due to that. But it was still
happening. What I found out is that if instead of using the full path
when creating a folder, we cd in the parent folder and create the
folder with just its name it works.