Reorganising Cypress tests
I've reorganised the files so that Cypress can be run locally as well as
with docker. The Test Runner is a more handy way to run the test while
writing new one and it requires Cypress to be install locally.

I've also started writing a README.md file to explain how to use it.

I've also added a more complex test which create a Vault.