Add Error Handling on the Broker
Now, any Controller Exception should show in the broker logs.
I also implemented it so that this exception is sent to the web app as
an ApiError object. At the moment this object only contains the
Exception but late rit could contain a message to display to the user or
anything else.
THe Exception will also get printed in the Web logs and show on the
error page.