Off ticket tweak to fix the docker deploy on my IMac
A couple of days ago the docker deploy stopped working on my machine for some reason.  I found stuff on the web that suggested that Alpine doesn't work with /bin/bash but does with /bin/sh.
I changed all the docker deployment scripts to use this and it started working again. I dunno why it suddenly changed or why it ever worked!

If this causes issues elsewhere we can easily revert.