IssueID #RSS212-69 First pass at fixing the restart function.
I've been able to restart jobs that fail just before the upload section starts. Deposits still work too!  It turns out that all the info we need to id the location of the failed files and the VM they were run on is in the db afterall.

I've checksummed the restarted then retrieved file against the orig and both still match.

I haven't done anything yet to ensure we go to the correct VM all testing so far has been on a docker deploy with only 1 worker.

1) Updated Receiver.java so that if the deposit is a restarted one we use the temp file location from the previous attempt instead of creating a new one.
I also updated the file so that we no longer clean up for failed jobs (as we need the files for any restart attempt)
2) Updated the broker DepositsController so that all the info for the packaging stage is added for a restarted job.  I also fixed a bug where the second filePath wasn't being added to the database.
3) Updated the Deposit task so that if the job is a restart job and the last event is one prior to the end of packaging that we redo the packaging if it is after we skip to the upload section.  This needs extended to skip upload / skip validation etc. but seems to work so far.

I also realised that as the restarts will be a manual admin process we already have an ugly hack for ensuring that the restarted job runs on the correct worker.  We could just kill all the unused workers before restarting then it would run on the correct one (I said it was ugly).

Hopefully I'll manage something a bit more elegant :-) Just in case I don't all the events for a deposit contain the VM in the agent value, it actually is process id @ hostname but we can easily strip out the hostname to leave only that worker up.