IssueID #RSS212-125 Make create vault the landing page if user has no vaults
1) Updated VaultsController so that getVaultsListing redirects to createVault if the user has no current vaults.

This kind of works at the moment the actual landing page is fine you see the right one depending on whether you have vaults or not.  If you see the create vault page as the landing and create a vault you see a permission denied box after submitting refreshing the page makes it go away but obviously that'll need fixed.