IssueID #RSS212-85 Fix calculation of policy status
1) Updated RetentionPoliciesService so that it only returns the review status if the check date is before now (like it was before)

I've tested on my machine by creating a new vault with a month retention date which showed the correct status.  It didn't fix any broken ones either when the page loaded or after the reviews scheduled job ran so it looks like any already broken ones will need to be fixed manually in the db (somehow).