IssueID #RSS212-147 Start of adding save functionality
I spent at bit of time working out how other views handle a partial save.  I eventually discovered that they just use the value param of submit buttons to id which button was clicked then a bit of JS to set the value as an hidden param to the controller.

The controller method then uses the param to decide if just the partial save or full add functionality is required.

In the current code nothing is saved but the logging shows the correct button has been pressed and returns the user to the form,  I'll finally start to add code to save to the database next and work on tidying up the code that returns to the form so that it goes back to the correct page rather than the first page.  At the moment you only briefly see the pending page before it goes back to the first page of the form.

1) Updated all of the save buttons in the various fieldset ftl files to be submit buttons and set a value which will be used to identify the type of button in the action
2) Did the same for the confirm button on the summary page
3) Updated CreatePrototype.ftl so that it has the required hidden values for spring security and to pass the button type to the action
4) Updated CreatePrototype.ftl so that the form points at the new prototype action
5) Updated CreatePrototype.ftl so that it now has a js function to set the submitAction param if the clicked button is a submit button
6) Updated VaultsController commented out the existing addVault and what now looks like abandoned savePartialVault controller
7) Updated VaultsController add new addVault which has the beginnings of the save or add functionality