IssueID #RSS212-147 Good bit of progress on saving a vault mid creation (still along way to go)
I've added a new PendingVaults table to the database like William proposed in a previous ticket.  The idea is that this table (which may or may not evolve to multiple tables) will store everything needed for the Vault to be curated and accepted.  Once the vault is curated it will be added to the usual Vault tables and removed from the pending table (So the pending table will only ever contain vaults in the process of being completed or accepted).

1) Added Bare bones PendingVaults model class
2) Added Bare bones PendingVaultsService, PendingVaultsDAO & PendingVaultsDAOImpl
3) Added code to the affirmation and info fieldsets files so that that the affirmation, vault name & vault description values bind the the model
4) Updated the VaultsControllers so that they now save these test values when the save button is clicked
5) Updated datavault-broker-root.xml changed the hibernate level back to vaildate & added config for new pendingVaultDAO class
6) Updated datavault-broker-servlet.xml added updated config for VaultsController to include the pendingVaultsService