IssueID #RSS212-147 Continued to add fields from the new fieldset interface to the database
I've made a start on the billing page values being added to the table.  So far the type of billing and the slice id are saved

I also added enums for the estimate and budget type radio button values.

Going to work on saved data prepopulating the form now before adding any more data to the pending vaults table.