IssueID #RSS212-153 New Oracle retrieve / delete command ready for testing
1) OracleObjectStorageClassic updated so that gen 2 versions of the retrieve / delete commands are available.

I've been able to test the delete one locally but not the retrieve one (as the current config never retrieves from Oracle).

I'll see if I can come up with some way of hacking things to use the new retrieve either that or deploy a temp Oracle only version.

Next I'll be changing the config to use an archive container like all our existing 'real' containers if that goes well along with the retrieval testing we should be ready to test on demo and hopefully finish this.

(though once we test on demo it might cause issues if we have to go back to the gen1 stuff on demo for whatever reason)