IssueID #RSS212-44 Started to add admin edit functionality to vault name
1) Updated vault.ftl so that in the summary of vault metadata section there is an edit button that triggers a popup and save button.
2) Updated web / broker VaultsControllers to have new updateVaultName method
3) Updated RestService to have new updateVaultName method

At the moment everything works like the edit description i.e. only people who have permission to edit the metadata can edit,  I'll need to change that so that it is only admins and to hide the button for people who don't have permission

When logged in as the correct user I have been able to edit the name (and description) but when I tried as an admin on someone else's vault it gave a permission error.