IssueID #RSS212-189 Started to persist the remaining new metadata values (NDMs, depositors, Owner, Pure Contact)
1) Updated CreateVault so it now contains new membera / accessors for isOwner, vaultOwner, nominatedDataManagers, depositors and pureContact
2) Updated usersFieldset.ftl to use the new CreateVault values
3) Updated VaultsController.addPendingVault to extract the new form values and made a start on add the pending roles (before getting sidetracked by how to do that exactly)
4) Updated RoleUtils so that there is now two new special roles that are created by default VaultCreator and NominatedDataManager (at least one more will follow),  I've made VaultCreator be added to the db with the same permissions as DataOwner but NDM is a placeholder at the moment.
5) Updated RoleDAO / RoleDAOImpl to have new getVaultCreator method
6) Updated RolesAndPermissionsService to have new getVaultCreator method
7) Updated VaultController.getPendingVaults method to use the new VaultCreator role to grant access (VaultCreator is basically the same role for pending vault as data owner is for vault)

Now that I have added the new VaultCreator role I am hoping that I can use the same table for pending and full vault roles and on acceptance they will just be updated so the pending vault id is removed and the vault id is added.

This is still a work in progress but I appear to be be able to see new pending vaults I create with the new Role and after manaully updating the role on my exisiting PVs can see those too.