IssueID #RSS212-210 Fixed bugs introduced on the various admin pages after the introduction of the new metadata style vaults
A number of the admin options were giving various errors due to new vaults no longer having a dataset from Pure.  I've added code to check if it exists and replace with "" this may or may not be the final fix depending on whether we decide to upgrade the original style vaults to new ones or have two different styles of vault.