IssueID #RSS212-130 Updated hasProducedPureRecord to user Vaults snapshot column to descide if it has produced it's pure record or not
1) Updated VaultsPureService.hasProducedPureRecord so it uses the vaults snapshot column to define whether the record has been produced or not
2) Updated CheckForPure.checkAll to fix a typo in the logging

So the plan for this is for all Vaults that haven't produced a snapshot to produce one sent to Damon an co then finally to update the Vaults entry in the db to have the snapshot.  We will use something else to confirm it has been added to Pure (Project ID possibly)

All the vaults created the old way will already have a value for snapshot and project id ( I think).

I'm going add the code to produce the XML next